initial commit
Funny enough, the CPU used in the ts7200 is a ARM920T, which uses the
ARMv4T architecture, the _same_ architecture that both the GBA and
original clickwheel iPods use!

As such, I took the foundations of a iPod clickwheel emulator I was
working on, and repurposed it as the foundation for this project! What
are the odds?