start of gdb stub + remove old debugger
You know, a small part of me is gonna miss my old janky debugger. That
thing proved to be surprisingly useful...

This commit undoes some stuff from the last commit, once again
reaffirming the age old adage that you shouldn't code while tired. I
mean, it's not like it was all bad, but there were definately a couple
places in the code that I was like "what the heck was I thinking??"

I'm going to try and keep the GDB stub interface fairly generic, since
it'd be nice to publish it as it's own crate (if I ever get to that
point). Unlike some other Rust GDB stubs I stumbled upon while browsing
GitHub, I'm trying to keep this one as "unintrusive" as possible. i.e:
instead of owning the target system, it only takes a reference to it.
This is nice, since it should allow starting / stopping a debugging
session somewhat dynamically. Moreover, it allows the target to have a
non-instrumented "step" function for when debugging is disabled, which
should be a boon to performance.