Extract Device trait from Memory trait
This also gave me an excuse to rework the whole "id" system. I've landed
on something that's _better_ (i.e: instead of moiving strings around,
a probe returns a singly-linked list of device identifiers, which is
serialized once at the end), but I think there's room for even _more_
generalization. This is good for now though.